jetstream_wireformat::fixed

Trait FixedWireFormat

pub trait FixedWireFormat: WireFormat {
    const WIRE_SIZE: usize;
}

A WireFormat type whose encoded size is known at compile time.

Required Associated Constants

const WIRE_SIZE: usize

Number of bytes every value of this type occupies on the wire.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".