Crate jetstream_wireformat

Modules

miette

wire_format_extensions

Structs

Data

A type that encodes an arbitrary number of bytes of data. Typically used for Rread Twrite messages. This differs from a Vec<u8> in that it encodes the number of bytes using a u32 instead of a u16.

P9String

A 9P protocol string.

Wrapped

Traits

WireFormat

A type that can be encoded on the wire using the 9P protocol.

Derive Macros

JetStreamWireFormat

Derives wire format encoding for structs